+++ src/App.js
@@ -1,12 +1,16 @@
 import React from 'react'
-import { Login} from './component'
+// import { NavBar} from './component'
+// import { Account} from './component'
+import { RoyalCourtNotices} from './component'
+// import { History } from './component'
 // import { PageTemplate } from './component'
 
 function App() {
   return (<>
-     {/* <PageTemplate header={}> */}
-      {/* <HistoryPage /> */}
-      <Login />
+     {/* <PageTemplate header={<NavBar />}> */}
+      {/* <History /> */}
+      {/* <Account /> */}
+      <RoyalCourtNotices />
     
      {/* </PageTemplate> */}
     </>
