+++ src/component/atoms/Input/index.css
@@ -1,0 +1,8 @@
+.input{
+    width: 90%;
+    border: 1px solid rgb(195, 195, 195);
+    border-radius: 50px;
+    padding: 14px;
+    margin-bottom: 1.5rem;
+    font-size: 0.9rem;
+}
