+++ src/component/atoms/PrevButton/index.css
@@ -2,7 +2,7 @@
     width:28px;
     height: 28px;
     border-radius: 100%;
-    /* border: 1px solid black; */
+    border: 1px solid black;
     position: absolute;
     top:24px;
     left:24px;
