+++ src/component/molecules/LoginForm/index.css
@@ -2,14 +2,7 @@
     display: block;
 }
 
-.login-form input {
-    width: 90%;
-    border: 1px solid rgb(195, 195, 195);
-    border-radius: 50px;
-    padding: 14px;
-    margin-bottom: 1.5rem;
-    font-size: 0.9rem;
-}
+
 
 .login-btn {
     background-color: rgb(210, 18, 32);
