+++ src/component/organisms/NavBar/index.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import { PrevButton } from '../../atoms'
+import { MobileMenuIcon } from '../MobileMenuIcon'
+
+export const NavBar = () => {
+  return (
+    <div>
+      <PrevButton />
+      <MobileMenuIcon />
+    </div>
+  )
+}
+
+
