+++ src/component/organisms/index.js
@@ -1,1 +1,2 @@
 export * from './MobileMenuIcon'
+export * from './NavBar'
