+++ src/component/pages/History/kingCrestData.js
@@ -1,0 +1,36 @@
+export const kingCrestData=[
+    {img:'https://atuwatseiii.com/assets/img/crests/iwere_kings/atonrongboye.jpg'},
+    {img:'https://atuwatseiii.com/assets/img/crests/iwere_kings/esigie.jpg'},
+    {img:'https://atuwatseiii.com/assets/img/crests/iwere_kings/ojoluwa.jpg'},
+    {img:'https://atuwatseiii.com/assets/img/crests/iwere_kings/irame.jpg'},
+    {img:'https://atuwatseiii.com/assets/img/crests/iwere_kings/ijijen.jpg'},
+    {img:'https://atuwatseiii.com/assets/img/crests/iwere_kings/ginuwa_i.jpg'},
+    {img:'https://atuwatseiii.com/assets/img/crests/iwere_kings/atuwatse_iii.jpg'},
+    {img:'https://atuwatseiii.com/assets/img/crests/iwere_kings/ikenwoli.jpg'},
+    {img:'https://atuwatseiii.com/assets/img/crests/iwere_kings/atuwatse_ii.jpg'},
+    {img:'https://atuwatseiii.com/assets/img/crests/iwere_kings/erejuwa_ii.jpg'},
+    {img:'https://atuwatseiii.com/assets/img/crests/iwere_kings/ginuwa_ii.jpg'},
+    {img:'https://atuwatseiii.com/assets/img/crests/iwere_kings/interregnum.jpg'},
+    {img:'https://atuwatseiii.com/assets/img/crests/iwere_kings/akengbuwa.jpg'},
+    {img:'https://atuwatseiii.com/assets/img/crests/iwere_kings/erejuwa.jpg'},
+    {img:'https://atuwatseiii.com/assets/img/crests/iwere_kings/atogbuwa.jpg'},
+    {img:'https://atuwatseiii.com/assets/img/crests/iwere_kings/akenjioye.jpg'},
+    {img:'https://atuwatseiii.com/assets/img/crests/iwere_kings/omagboye.jpg'},
+    {img:'https://atuwatseiii.com/assets/img/crests/iwere_kings/akenjioye_sebastien.jpg'},
+    {img:'https://atuwatseiii.com/assets/img/crests/iwere_kings/abejeoye.jpg'},
+    {img:'https://atuwatseiii.com/assets/img/crests/iwere_kings/omoluyiri.jpg'},
+    {img:'https://atuwatseiii.com/assets/img/crests/iwere_kings/oyenakpagha.jpg'},
+    {img:'https://atuwatseiii.com/assets/img/crests/iwere_kings/atuwatse.jpg'},
+]
+
+// let counter =0
+// const increaseCounter= () => {
+//   const getCounter=()=> counter
+//   counter =+ 1
+//   getCounter()
+//   return counter
+// }
+// setInterval(() => {
+//   increaseCounter()
+//   console.log(increaseCounter())
+//   }, 1000);
