--- src/component/pages/HistoryPage/index.css
@@ -1,70 +1,0 @@
-.history-body-section{
-    display: grid;
-    grid-template-columns: 40% 60%;
-}
-.king-crest-container{
-    /* height: 100vh; */
-    background-color: white;
-    padding: 18vh 0 0 5vw;
-}
-.king-crest-container div{
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    width: 90%;
-}
-.mob-history-text {
-    display: none;
-    text-align: left;
-    margin: 0;
-}
-.king-crest-container button{
-    background-color: transparent;
-    border: 0.12rem solid black;
-    border-radius: 6px;
-    width: 22px;
-    height: 35px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    cursor: pointer;
-}
-
-
-.king-crest-container img {
-    width: 100%;
-    height: 100%;
-    padding-bottom: 10px;
-}
-.history-page-right-half{
-    padding: 24px;
-}
-.history-page-right-half h2{
-    font-size: 2rem;
-}
-
-/* Mobile View */
-@media only screen and (max-width: 900px) {
-    .history-body-section {
-        display: block;
-    }
-    .king-crest-container{
-        /* height: 70vh; */
-        padding: 9vh 0 0 5vw;
-    }
-    .mob-history-text {
-        display: block;
-    }
-    .desk-history-text{
-        display: none;
-    }
-}
-/* Tab view */
-@media only screen and (min-width: 768px) {
-    .king-crest-img-container{
-        /* width: 50vw !important; */
-        padding: 0%;
-        margin: 0;
-    }
-
-}
