+++ src/index.css
@@ -9,8 +9,8 @@ h1,h2{
 h1,h2,h3,h4,h5,h6{
   line-height: 1.2;
 }
-@media only screen and (max-width: 900px) {
+/* @media only screen and (max-width: 900px) {
   body {
     background-color: white;
   }
-}
+} */
